rename client to tinfoilai
