+++ src/tinfoilai.ts
